Update Comment controller and service

+++ src/main/java/com/app/Service/CommentService.java
@@ -3,9 +3,13 @@ package com.app.Service;
 import com.app.DAO.Impl.CommentDaoImpl;
 import com.app.DTO.CommentDTO;
 import com.app.Model.Comment;
+import com.app.Model.User;
+import com.app.exception.sub.UserNotMatchException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 public class CommentService {
@@ -24,10 +28,12 @@ public class CommentService {
         return commentDao.save(comment);
     }
 
-    public void deleteById(Long id){
+    public void deleteIfUserMatches(User user, Long id){
         Comment comment = commentDao.findById(id);
-        if(comment != null){
+        if(comment != null && Objects.equals(user.getId(),comment.getUser().getId())){
             commentDao.deleteById(comment.getId());
+        } else {
+            throw new UserNotMatchException("User does not match with a comment");
         }
     }
 }
